@@ -13,38 +13,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.8",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.8",
-    "@angular/compiler": "^6.1.8",
-    "@angular/core": "^6.1.8",
-    "@angular/forms": "^6.1.8",
-    "@angular/http": "^6.1.8",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^6.1.8",
-    "@angular/platform-browser-dynamic": "^6.1.8",
-    "@angular/router": "^6.1.8",
-    "@covalent/core": "2.0.0-beta.3",
-    "@ngrx/effects": "^6.1.0",
-    "@ngrx/entity": "^6.1.0",
-    "@ngrx/router-store": "^6.1.0",
-    "@ngrx/store": "^6.1.0",
-    "@ngrx/store-devtools": "^6.1.0",
+    "@angular/animations": "^7.2.3",
+    "@angular/cdk": "^7.3.1",
+    "@angular/common": "^7.2.3",
+    "@angular/compiler": "^7.2.3",
+    "@angular/core": "^7.2.3",
+    "@angular/forms": "^7.2.3",
+    "@angular/http": "^7.2.3",
+    "@angular/material": "^7.3.1",
+    "@angular/platform-browser": "^7.2.3",
+    "@angular/platform-browser-dynamic": "^7.2.3",
+    "@angular/router": "^7.2.3",
+    "@covalent/core": "^2.0.1",
+    "@ngrx/effects": "^7.2.0",
+    "@ngrx/entity": "^7.2.0",
+    "@ngrx/router-store": "^7.2.0",
+    "@ngrx/store": "^7.2.0",
+    "@ngrx/store-devtools": "^7.2.0",
     "@types/lodash": "^4.14.104",
+    "ag-grid-angular": "^21.0.1",
+    "ag-grid-community": "^21.0.1",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
     "lodash": "^4.17.5",
     "moment": "^2.20.1",
     "ngrx-store-freeze": "^0.2.1",
-    "rxjs": "^6.3.2",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.6",
-    "@angular/cli": "^6.2.3",
-    "@angular/compiler-cli": "^6.1.8",
-    "@angular/language-service": "^6.1.8",
-    "@ngrx/schematics": "^6.1.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.0",
+    "@angular/compiler-cli": "^7.2.3",
+    "@angular/language-service": "^7.2.3",
+    "@ngrx/schematics": "^7.2.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -60,6 +63,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.9.2"
+    "typescript": "~3.2.4"
   }
 }
