@@ -1,30 +1,20 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
 
-import { Store, select } from '@ngrx/store';
-import * as fromStore from '../reducers/application.reducre';
+import { Store } from '@ngrx/store';
+import * as fromStore from '../reducers/application.reducer';
 import * as fromApplication from '../actions/application.actions';
 
 import { Actions, Effect, ofType } from '@ngrx/effects';
 
-import {
-  LoadSucursal,
-  ApplicationActionTypes,
-  LoadSucursalSuccess,
-  LoadSucursalFail
-} from '../actions/application.actions';
-
 import { Observable, defer, of } from 'rxjs';
-import { tap, map, mergeMap, catchError } from 'rxjs/operators';
+import { tap, map } from 'rxjs/operators';
 
-import { Sucursal } from '../../models';
 import {
   Router,
   NavigationStart,
   NavigationEnd,
   NavigationError,
-  NavigationCancel,
-  RoutesRecognized
+  NavigationCancel
 } from '@angular/router';
 import {
   ROUTER_NAVIGATION,
@@ -32,14 +22,16 @@ import {
   RouterCancelAction,
   ROUTER_CANCEL
 } from '@ngrx/router-store';
+import { TdLoadingService, TdDialogService } from '@covalent/core';
 
 @Injectable()
 export class ApplicationsEffects {
   constructor(
     private actions$: Actions,
-    private http: HttpClient,
     private router: Router,
-    private store: Store<fromStore.State>
+    private store: Store<fromStore.State>,
+    private loadingService: TdLoadingService,
+    private dialogService: TdDialogService
   ) {
     this.router.events.subscribe(event => {
       switch (true) {
@@ -63,21 +55,39 @@ export class ApplicationsEffects {
       }
     });
   }
-  /*
-  @Effect()
-  loadSucursal$ = this.actions$.pipe(
-    ofType<LoadSucursal>(ApplicationActionTypes.LoadSucursal),
-    map(action => `${action.payload}/sucursal`),
-    mergeMap(url => {
-      return this.http
-        .get<Sucursal>(url)
-        .pipe(
-          map(sucursal => new LoadSucursalSuccess({ sucursal })),
-          catchError(error => of(new LoadSucursalFail(error)))
-        );
+
+  @Effect({ dispatch: false })
+  loading$ = this.actions$.pipe(
+    ofType<fromApplication.SetGlobalLoading>(
+      fromApplication.ApplicationActionTypes.SetGlobalLoading
+    ),
+    tap(action => {
+      if (action.payload.loading) {
+        this.loadingService.register();
+      } else {
+        this.loadingService.resolve();
+      }
     })
   );
-  */
+
+  @Effect({ dispatch: false })
+  errorHandler$ = this.actions$.pipe(
+    ofType<fromApplication.GlobalHttpError>(
+      fromApplication.ApplicationActionTypes.GlobalHttpError
+    ),
+    map(action => action.payload.response),
+    map(response => {
+      const message = response.error ? response.error.message : 'Error';
+      const message2 = response.message ? response.message : '';
+      console.error('Error: ', response);
+      this.dialogService.openAlert({
+        message: `${response.status} ${message} ${message2}`,
+        title: `Error ${response.status}`,
+        closeButton: 'Cerrar'
+      });
+    })
+  );
+
   @Effect({ dispatch: false })
   init$: Observable<any> = defer(() => of(null)).pipe(
     tap(() => console.log('Effect inicial de la applicacion: init$'))
