@@ -4,7 +4,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { MaterialModule } from './_material/material.module';
 import { CovalentModule } from './_covalent/covalent.module';
-
+import { AgGridModule } from 'ag-grid-angular';
 // components
 import { components, entyComponents } from './components';
 // Directives
@@ -18,7 +18,8 @@ import { MAT_DATE_LOCALE } from '@angular/material';
     FormsModule,
     ReactiveFormsModule,
     MaterialModule,
-    CovalentModule
+    CovalentModule,
+    AgGridModule.withComponents([])
   ],
   exports: [
     CommonModule,
@@ -26,6 +27,7 @@ import { MAT_DATE_LOCALE } from '@angular/material';
     ReactiveFormsModule,
     MaterialModule,
     CovalentModule,
+    AgGridModule,
     ...components,
     ...entyComponents,
     ...directives,
