@@ -5,7 +5,7 @@ import { of } from 'rxjs';
 
 import * as fromClases from '../actions/clases.actions';
 
-import { Effect, Actions } from '@ngrx/effects';
+import { Effect, Actions, ofType } from '@ngrx/effects';
 import { ClasesService } from '../../services';
 
 @Injectable()
@@ -13,40 +13,13 @@ export class ClasesEffects {
   constructor(private actions$: Actions, private service: ClasesService) {}
 
   @Effect()
-  loadClases$ = this.actions$.ofType(fromClases.LOAD_CLASES).pipe(
+  loadClases$ = this.actions$.pipe(
+    ofType(fromClases.LOAD_CLASES),
     switchMap(() => {
-      return this.service
-        .list()
-        .pipe(
-          map(clases => new fromClases.LoadClasesSuccess(clases)),
-          catchError(error => of(new fromClases.LoadClasesFail(error)))
-        );
-    })
-  );
-
-  @Effect()
-  createClase$ = this.actions$.ofType(fromClases.CREATE_CLASE).pipe(
-    map((action: fromClases.CreateClase) => action.payload),
-    switchMap(clase => {
-      return this.service
-        .save(clase)
-        .pipe(
-          map(res => new fromClases.CreateClaseSuccess(res)),
-          catchError(error => of(new fromClases.CreateClaseFail(error)))
-        );
-    })
-  );
-
-  @Effect()
-  updateClase$ = this.actions$.ofType(fromClases.UPDATE_CLASE).pipe(
-    map((action: fromClases.UpdateClase) => action.payload),
-    switchMap(clase => {
-      return this.service
-        .update(clase)
-        .pipe(
-          map(res => new fromClases.UpdateClaseSuccess(res)),
-          catchError(error => of(new fromClases.UpdateClaseFail(error)))
-        );
+      return this.service.list().pipe(
+        map(clases => new fromClases.LoadClasesSuccess(clases)),
+        catchError(error => of(new fromClases.LoadClasesFail(error)))
+      );
     })
   );
 }
